added prototype code + setup for multilevel space routes; moved files around
