+++ src/domain/collaboration/whiteboard/WhiteboardDialog/WhiteboardDialogFooter.tsx
@@ -7,7 +7,7 @@ import { Actions } from '@/core/ui/actions/Actions';
 import { Trans, useTranslation } from 'react-i18next';
 import { useAuthenticationContext } from '@/core/auth/authentication/hooks/useAuthenticationContext';
 import { CommunityMembershipStatus, ContentUpdatePolicy, SpaceLevel } from '@/core/apollo/generated/graphql-schema';
-import { useSpace } from '@/domain/journey/space/SpaceContext/useSpace';
+import { useSpace } from '@/domain/space/SpaceContext/useSpace';
 import { useSubSpace } from '@/domain/journey/subspace/hooks/useSubSpace';
 import RouterLink from '@/core/ui/link/RouterLink';
 import { buildLoginUrl } from '@/main/routing/urlBuilders';
