+++ src/domain/community/inviteContributors/InviteVCsDialog.tsx
@@ -6,7 +6,7 @@ import { AiPersonaBodyOfKnowledgeType, RoleName, RoleSetContributorType } from '
 import DialogWithGrid from '@/core/ui/dialog/DialogWithGrid';
 import Gutters from '@/core/ui/grid/Gutters';
 import DialogHeader from '@/core/ui/dialog/DialogHeader';
-import { useSpace } from '@/domain/journey/space/SpaceContext/useSpace';
+import { useSpace } from '@/domain/space/SpaceContext/useSpace';
 import useInviteContributors from '@/domain/access/_removeMe/useInviteContributors';
 import VCIcon from '@/domain/community/virtualContributor/VirtualContributorsIcons';
 import { ContributorProps, InviteContributorDialogProps } from './InviteContributorsProps';
