+++ src/domain/journey/opportunity/pages/OpportunitySettings/OpportunitySettingsView.tsx
@@ -18,7 +18,7 @@ import DeleteIcon from '@/domain/journey/space/pages/SpaceSettings/icon/DeleteIc
 import EntityConfirmDeleteDialog from '@/domain/journey/space/pages/SpaceSettings/EntityConfirmDeleteDialog';
 import { useSubSpace } from '@/domain/journey/subspace/hooks/useSubSpace';
 import PageContent from '@/core/ui/content/PageContent';
-import { useSpace } from '@/domain/journey/space/SpaceContext/useSpace';
+import { useSpace } from '@/domain/space/SpaceContext/useSpace';
 
 const OpportunitySettingsView = () => {
   const { t } = useTranslation();
