+++ src/domain/journey/space/templates/SpaceTemplatesQueries.graphql
