+++ src/domain/space/layout/TabbedSpaceL0/Tabs/SpaceSubspaces/SubspaceListView.tsx
@@ -16,7 +16,7 @@ import {
   useUpdateTemplateDefaultMutation,
 } from '@/core/apollo/generated/apollo-hooks';
 import { useNotification } from '@/core/ui/notifications/useNotification';
-import { useSpace } from '@/domain/journey/space/SpaceContext/useSpace';
+import { useSpace } from '@/domain/space/SpaceContext/useSpace';
 import { JourneyCreationDialog } from '@/domain/shared/components/JourneyCreationDialog/JourneyCreationDialog';
 import SubspaceIcon2 from '@/domain/journey/subspace/icon/SubspaceIcon2';
 import { JourneyFormValues } from '@/domain/shared/components/JourneyCreationDialog/JourneyCreationForm';
